requirements.txt
flet>=0.23.0
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0
